Update digital ty processing page to include opt in
