+++ support-frontend/assets/pages/digital-subscription-checkout/components/thankYou/thankYouProducts.jsx
@@ -8,7 +8,7 @@ import Content from 'components/content/content';
 import Text, { LargeParagraph } from 'components/text/text';
 import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
 
-import AppsSection from './components/thankYou/appsSection';
+import AppsSection from './appsSection';
 
 // ----- Types ----- //
 
@@ -18,7 +18,7 @@ type PropTypes = {
 
 // ----- Component ----- //
 
-function ThankYouExistingContent({ countryGroupId }: PropTypes) {
+function ThankYouProducts({ countryGroupId }: PropTypes) {
 
   return (
     <div>
@@ -37,4 +37,4 @@ function ThankYouExistingContent({ countryGroupId }: PropTypes) {
 
 // ----- Export ----- //
 
-export default ThankYouExistingContent;
+export default ThankYouProducts;
