+++ support-frontend/assets/pages/digital-subscription-checkout/thankYouExisting.jsx
@@ -17,12 +17,8 @@ import SubscriptionTermsPrivacy from 'components/legal/subscriptionTermsPrivacy/
 import SubscriptionFaq from 'components/subscriptionFaq/subscriptionFaq';
 import 'stylesheets/skeleton/skeleton.scss';
 
-import HeadingBlock from 'components/headingBlock/headingBlock';
-import { HeroWrapper } from 'components/productPage/productPageHero/productPageHero';
-
 import ReturnSection from 'components/subscriptionCheckouts/thankYou/returnSection';
-import ThankYouExistingContent from './thankYouExistingContent';
-import ThankYouContent from './components/thankYou/hero';
+import ThankYouContent from './components/thankYou/thankYouContent';
 
 import './digitalSubscriptionCheckout.scss';
 
@@ -45,16 +41,8 @@ const content = (
           <SubscriptionFaq subscriptionProduct="DigitalPack" />
         </Footer>}
     >
-      <div className="thank-you-stage">
-        <HeroWrapper appearance="custom">
-          <ThankYouContent countryGroupId={countryGroupId} />
-          <HeadingBlock>
-            Your Digital Subscription is already live
-          </HeadingBlock>
-        </HeroWrapper>
-        <ThankYouExistingContent countryGroupId={countryGroupId} />
-        <ReturnSection subscriptionProduct="DigitalPack" />
-      </div>
+      <ThankYouContent countryGroupId={countryGroupId} />
+      <ReturnSection subscriptionProduct="DigitalPack" />
     </Page>
   </Provider>
 );
